Improve logic of privileged_writable
Raise alerts in a number of extra situations:

1. If a low privileged user can write to a file there should always be a
warning, not just if YOU are the current owner. If the tool is being run
by root for auditing purposes then it's important to show if a standard
user can write to a script they're permitted to run with elevated
permissions.

2. Furthermore, if a low privileged user owns a privileged file then it
should be reported no matter what the current permissions are. The owner
can simply set the write permission if they wish.

3. Similarly to (1), if an untrusted group can write to a privileged
file then it should be a warning even if the current user isn't a member
of the group.